Ignore invalid code

+++ app/src/main/java/com/lidorttol/rickandmorty/data/remote/api/ApiService.kt
@@ -14,6 +14,6 @@ interface ApiService {
     suspend fun getCharacters(): CharacterRemoteResponseDto<CharacterDto>
 
     @GET("character/{id}")
-    suspend fun getCharacterById(@Path("id") id: Int) : CharacterDto
+    suspend fun getCharacterById(@Path("id") id: Long) : CharacterDto
 
 }
